requirements.txt
python-binance
pandas
plotly
matplotlib
PyYAML
requests
tzlocal
keras
neupy
scikit-learn
scipy
tensorflow==2.7.0
tensorflow-probability
tensorflow-datasets==4.1.0
Pillow==9.0.0
statsmodels
numpy
arch
